+++ scripts/internal/create_work_dir.sh
@@ -11,6 +11,7 @@ TARGET_FIRMWARE_PATH="$(cut -d "/" -f 1 -s <<< "$TARGET_FIRMWARE")_$(cut -d "/" 
 COPY_SOURCE_FIRMWARE()
 {
     local SOURCE_FOLDERS="product system"
+    $TARGET_USE_DYNAMIC_PARTITIONS || SOURCE_FOLDERS="odm product system prism optics"
     for f in $SOURCE_FOLDERS; do
         if [ -d "$FW_DIR/$SOURCE_FIRMWARE_PATH/$f" ]; then
             LOG "- Copying /$f from source firmware"
@@ -111,6 +112,7 @@ COPY_SOURCE_FIRMWARE()
 COPY_TARGET_FIRMWARE()
 {
     local TARGET_FOLDERS="odm odm_dlkm system_dlkm vendor vendor_dlkm"
+    $TARGET_USE_DYNAMIC_PARTITIONS || TARGET_FOLDERS="odm_dlkm system_dlkm vendor vendor_dlkm"
     for f in $TARGET_FOLDERS; do
         if [ -d "$FW_DIR/$TARGET_FIRMWARE_PATH/$f" ]; then
             LOG "- Copying /$f from target firmware"